scraper/requirements.txt
requests==2.31.0
gspread==6.1.4
google-auth==2.29.0
python-dotenv==1.0.1
pytest==8.2.0
